+++ src/lib.rs
@@ -67,6 +67,7 @@ use std::{
     fmt, io,
     ops::{Deref, DerefMut, Range},
     path::{Path, PathBuf},
+    sync::atomic::AtomicU64,
     sync::Arc,
     time::Duration,
 };
@@ -106,11 +107,14 @@ pub use libp2p::{
 };
 
 use libp2p::{
+    core::{muxing::StreamMuxerBox, transport::Boxed},
     kad::{store::MemoryStoreConfig, KademliaConfig, Mode},
     ping::Config as PingConfig,
     swarm::dial_opts::DialOpts,
     StreamProtocol,
 };
+
+pub(crate) static BITSWAP_ID: AtomicU64 = AtomicU64::new(1);
 
 #[derive(Debug, Clone)]
 pub enum StoragePath {
@@ -483,6 +487,15 @@ impl From<InnerConnectionEvent> for ConnectionEvent {
 
 type TSwarmEvent<C> = <TSwarm<C> as Stream>::Item;
 type TSwarmEventFn<C> = Arc<dyn Fn(&mut TSwarm<C>, &TSwarmEvent<C>) + Sync + Send>;
+type TTransportFn = Box<
+    dyn Fn(
+            &Keypair,
+            Option<libp2p::relay::client::Transport>,
+        ) -> std::io::Result<Boxed<(PeerId, StreamMuxerBox)>>
+        + Sync
+        + Send
+        + 'static,
+>;
 
 #[derive(Debug, Copy, Clone)]
 pub enum FDLimit {
@@ -491,7 +504,7 @@ pub enum FDLimit {
 }
 
 /// Configured Ipfs which can only be started.
-
+#[allow(clippy::type_complexity)]
 pub struct UninitializedIpfs<C: NetworkBehaviour<ToSwarm = void::Void> + Send> {
     keys: Keypair,
     options: IpfsOptions,
@@ -499,6 +512,7 @@ pub struct UninitializedIpfs<C: NetworkBehaviour<ToSwarm = void::Void> + Send> {
     delay: bool,
     swarm_event: Option<TSwarmEventFn<C>>,
     custom_behaviour: Option<C>,
+    custom_transport: Option<TTransportFn>,
 }
 
 pub type UninitializedIpfsNoop = UninitializedIpfs<libp2p::swarm::dummy::Behaviour>;
@@ -531,6 +545,7 @@ impl<C: NetworkBehaviour<ToSwarm = void::Void> + Send> UninitializedIpfs<C> {
             delay,
             swarm_event: None,
             custom_behaviour: None,
+            custom_transport: None,
         }
     }
 
@@ -677,8 +692,19 @@ impl<C: NetworkBehaviour<ToSwarm = void::Void> + Send> UninitializedIpfs<C> {
         self
     }
 
+    /// Set a custom behaviour
     pub fn set_custom_behaviour(mut self, behaviour: C) -> Self {
         self.custom_behaviour = Some(behaviour);
+        self
+    }
+
+    #[allow(clippy::type_complexity)]
+    /// Set a transport
+    pub fn set_custom_transport(
+        mut self,
+        transport: TTransportFn,
+    ) -> Self {
+        self.custom_transport = Some(transport);
         self
     }
 
@@ -713,6 +739,7 @@ impl<C: NetworkBehaviour<ToSwarm = void::Void> + Send> UninitializedIpfs<C> {
             mut options,
             swarm_event,
             custom_behaviour,
+            custom_transport,
             ..
         } = self;
 
@@ -829,7 +856,7 @@ impl<C: NetworkBehaviour<ToSwarm = void::Void> + Send> UninitializedIpfs<C> {
             transport_config,
             repo.clone(),
             exec_span,
-            custom_behaviour,
+            (custom_behaviour, custom_transport),
         )
         .instrument(tracing::trace_span!(parent: &init_span, "swarm"))
         .await?;
@@ -2173,6 +2200,7 @@ mod node {
         }
 
         /// Returns the subscriptions for a `Node`.
+        #[allow(clippy::type_complexity)]
         pub fn get_subscriptions(
             &self,
         ) -> &parking_lot::Mutex<HashMap<Cid, Vec<oneshot::Sender<Result<Block, String>>>>>