+++ src/p2p/behaviour.rs
@@ -581,6 +581,8 @@ impl<C: NetworkBehaviour<ToSwarm = void::Void>> Behaviour<C> {
             self.addressbook.add_address(peer, addr);
         }
 
+        self.pubsub.add_explicit_peer(&peer);
+
         // if let Some(bitswap) = self.bitswap.as_ref() {
         //     let client = bitswap.client().clone();
         //     let server = bitswap.server().cloned();