+++ src/p2p/mod.rs
@@ -6,7 +6,7 @@ use std::num::{NonZeroU8, NonZeroUsize};
 
 use crate::error::Error;
 use crate::repo::Repo;
-use crate::IpfsOptions;
+use crate::{IpfsOptions, TTransportFn};
 
 use either::Either;
 use libp2p::gossipsub::ValidationMode;
@@ -269,6 +269,7 @@ impl Default for SwarmConfig {
     }
 }
 
+#[allow(clippy::type_complexity)]
 /// Creates a new IPFS swarm.
 pub async fn create_swarm<C: NetworkBehaviour<ToSwarm = void::Void>>(
     keypair: &Keypair,
@@ -277,7 +278,7 @@ pub async fn create_swarm<C: NetworkBehaviour<ToSwarm = void::Void>>(
     transport_config: TransportConfig,
     repo: Repo,
     span: Span,
-    custom: Option<C>,
+    (custom, custom_transport): (Option<C>, Option<TTransportFn>),
 ) -> Result<TSwarm<C>, Error> {
     let keypair = keypair.clone();
     let peer_id = keypair.public().to_peer_id();
@@ -287,7 +288,10 @@ pub async fn create_swarm<C: NetworkBehaviour<ToSwarm = void::Void>>(
             .await?;
 
     // Set up an encrypted TCP transport over the Yamux and Mplex protocol. If relay transport is supplied, that will be apart
-    let transport = transport::build_transport(keypair, relay_transport, transport_config)?;
+    let transport = match custom_transport {
+        Some(transport) => transport(&keypair, relay_transport)?,
+        None => transport::build_transport(keypair, relay_transport, transport_config)?,
+    };
 
     // Create a Swarm
     let swarm = libp2p::swarm::SwarmBuilder::with_executor(