+++ src/repo/mod.rs
@@ -310,6 +310,7 @@ impl<C: Borrow<Cid>> PinKind<C> {
 ///
 /// Consolidates a blockstore, a datastore and a subscription registry.
 
+#[allow(clippy::type_complexity)]
 #[derive(Debug, Clone)]
 pub struct Repo {
     block_store: Arc<dyn BlockStore>,
@@ -473,16 +474,24 @@ impl Repo {
 
     /// Retrives a block from the block store, or starts fetching it from the network and awaits
     /// until it has been fetched.
+    #[inline]
     pub async fn get_block(
         &self,
         cid: &Cid,
         peers: &[PeerId],
         local_only: bool,
     ) -> Result<Block, Error> {
-        // FIXME: here's a race: block_store might give Ok(None) and we get to create our
-        // subscription after the put has completed. So maybe create the subscription first, then
-        // cancel it?
+        self.get_block_with_session(None, cid, peers, local_only)
+            .await
+    }
 
+    pub(crate) async fn get_block_with_session(
+        &self,
+        session: Option<u64>,
+        cid: &Cid,
+        peers: &[PeerId],
+        local_only: bool,
+    ) -> Result<Block, Error> {
         if let Some(block) = self.get_block_now(cid).await? {
             Ok(block)
         } else {
@@ -494,15 +503,11 @@ impl Repo {
 
             self.subscriptions.lock().entry(*cid).or_default().push(tx);
 
-            // let subscription = self
-            //     .subscriptions
-            //     .create_subscription((*cid).into(), Some(self.events.clone()));
-
             // sending only fails if no one is listening anymore
             // and that is okay with us.
             self.events
                 .clone()
-                .send(RepoEvent::WantBlock(None, *cid, peers.to_vec()))
+                .send(RepoEvent::WantBlock(session, *cid, peers.to_vec()))
                 .await
                 .ok();